+++ tests/e2e/utils/src/flows/merchant.js
@@ -23,6 +23,7 @@ const {
 	WP_ADMIN_NEW_SHIPPING_ZONE,
 	WP_ADMIN_ANALYTICS_PAGES,
 	WP_ADMIN_ALL_USERS_VIEW,
+	WP_ADMIN_IMPORT_PRODUCTS,
 	IS_RETEST_MODE,
 } = require( './constants' );
 
@@ -196,6 +197,12 @@ const merchant = {
 
 	openAllUsersView: async () => {
 		await page.goto( WP_ADMIN_ALL_USERS_VIEW, {
+			waitUntil: 'networkidle0',
+		} );
+	},
+
+  openImportProducts: async () => {
+		await page.goto( WP_ADMIN_IMPORT_PRODUCTS , {
 			waitUntil: 'networkidle0',
 		} );
 	},